Fix various printf formats for 32bit, 32bit w/_FILE_OFFSET_BITS=64 and 64-bit

+++ src/zm_image.cpp
@@ -914,7 +914,7 @@ bool Image::ReadRaw(const char *filename) {
 
   if ( (unsigned int)statbuf.st_size != size ) {
     fclose(infile);
-    Error("Raw file size mismatch, expected %d bytes, found %ld", size, statbuf.st_size);
+    Error("Raw file size mismatch, expected %d bytes, found %jd", size, static_cast<intmax_t>(statbuf.st_size));
     return false;
   }
